backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.128.5
greenlet==3.3.1
gunicorn==25.0.3
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.2
opencv-python==4.13.0.92
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
passlib==1.7.4
pillow==12.1.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
